requirements.txt
py-stackexchange==2.2.7
requests==2.28.2
pandas==1.5.3
scikit-learn==1.2.1